+++ src/glGribLoader.h
@@ -31,9 +31,9 @@ public:
   HandlePtr handleFromFile (const std::string &);
   void load (FieldFloatBufferPtr *, const std::vector<std::string> &, const OptionsGeometry &, float, FieldMetadata *, int = 1, int = 0, bool diff = false);
   void load (FieldFloatBufferPtr *, const std::string &, const OptionsGeometry &, FieldMetadata *);
-  void setSize (int _size) { size = _size; }
+  void setSize (size_t _size) { size = _size; }
 private:
-  int size = 0;
+  size_t size = 0;
 
   class entry_t
   {
